Add new functions for getting map data from tileservers
Adds functions for getting tiles and maps from tileservers, as well as
several different tileservers.
Parts of the functions are derived from the get_stamenmap function,
which essentially works the same as the new get_tilemap function, except
for a couple of key differences.
get_tilemap allows the user to specify custom tile servers, and thus use
any tileserver instead of only stamen's tileservers.
This also avoids an issue with get_stamenmap which fails expecting a jpg
but getting a png.