Implement guessing for zoom levels
When option zoom_guess is set to true, the get_tilemap functions will
now try to guess the zoom level by zooming in until min_tiles is
satisfied. Min_tiles can be either length one or two, describing either
the total amount of tiles needed, or the number of tiles in (x, y)
needed.

This allows the functions to be used in an automated fashion with no
prior knowledge about the map dimensions needed.